changed size required to the number on the checker

+++ 0x0B-malloc_free/100-argstostr.c
@@ -12,7 +12,7 @@
 char *argstostr(int ac, char **av)
 {
 	int i;
-	int check_len;
+	size_t check_len;
 	char *p_str;
 
 	check_len = 0;
